remove dependency on setuptools with a Python 3-only namespace package and "importlib.metadata"
Replacing "pkg_resources" with built-in "importlib.metadata" seems like a good idea (better
maintainance, less additional packages).

Then namespace package needs setuptools and we can just rely on native namespace packages
for Python 3 in that case. Compatibility with other "schwarz.*" packages is not a concern
because this repo contains a standalone application which is unlikely to be combined with
other "schwarz.*" packages.